--- src/components/left-side/OverviewStats/OverviewStatsTabList.svelte
@@ -1,11 +1,0 @@
-<div class="TabList">
-	<slot />
-</div>
-
-<style>
-	.TabList {
-		display: table;
-		width: 100%;
-		table-layout: fixed;
-	}
-</style>
