+++ src/components/left-side/OverviewStats/index.ts
@@ -1,7 +1,5 @@
 export { default as MoreButton } from './MoreButton.svelte';
 export { default as OverviewStatsBox } from './OverviewStatsBox.svelte';
 export { default as OverviewStatsTab } from './OverviewStatsTab.svelte';
-export { default as OverviewStatsTabList } from './OverviewStatsTabList.svelte';
 export { default as OverviewStatsTabPanel } from './OverviewStatsTabPanel.svelte';
 export { default as OverviewStatsTabPanelItem } from './OverviewStatsTabPanelItem.svelte';
-export { default as OverviewStatsTabs } from './OverviewStatsTabs.svelte';
